# Conflicts:
#	app/src/com/constellio/app/entities/batchProcess/ChangeValueOfMetadataBatchAsyncTask.java
#	app/src/com/constellio/app/modules/rm/ConstellioRMModule.java
#	app/src/com/constellio/app/modules/rm/ui/pages/folder/DisplayFolderViewImpl.java
#	app/src/com/constellio/app/ui/framework/data/RecordTextInputDataProvider.java
#	app/src/com/constellio/app/ui/pages/search/SearchPresenter.java
#	model/src/com/constellio/model/services/contents/ContentManager.java
#	resources_i18n/core/schemasManagementViews.properties
#	resources_i18n/core/security.properties
#	resources_i18n/i18n.properties
#	resources_i18n/i18n_en.properties
#	resources_i18n/rm/foldersAndDocuments.properties
#	resources_i18n/rm/foldersAndDocuments_en.properties
#	sdk/src/com/constellio/model/services/contents/ContentManagerScanAcceptanceTest.java

+++ model/src/com/constellio/model/services/contents/ContentManager.java
@@ -199,8 +199,8 @@ public class ContentManager implements StatefulService {
 					try {
 						createContentScanLockFile();
 						VaultScanResults vaultScanResults = new VaultScanResults();
-						scanVaultContentAndDeleteUnreferencedFiles(vaultScanResults);
 						addReportsAsTemporaryRecords(vaultScanResults);
+						scanVaultContentAndDeleteUnreferencedFiles(vaultScanResults);
 					} catch (IOException e) {
 						e.printStackTrace();
 					} catch (RecordServicesException e) {
@@ -296,13 +296,17 @@ public class ContentManager implements StatefulService {
 			if (!allReferencedHashes.contains(fileId)) {
 				File file = contentDao.getFileOf(fileId);
 				if (file.exists()) {
-					try {
-						contentDao.delete(asList(fileId, fileId + "__parsed", fileId + ".preview"));
-						vaultScanResults.incrementNumberOfDeletedContents();
-						vaultScanResults.appendMessage("INFO: Successfully deleted file " + file.getName() + "\n");
-					} catch (Exception e) {
-						vaultScanResults.appendMessage("ERROR: Could not delete file " + file.getName() + "\n");
-						vaultScanResults.appendMessage(e.getMessage() + "\n");
+					LocalDateTime lastModification = contentDao.getLastModification(fileId);
+					if (lastModification.isBefore(TimeProvider.getLocalDateTime().minus(configuration.getDelayBeforeDeletingUnreferencedContents()))) {
+						try {
+							contentDao.delete(asList(fileId, fileId + "__parsed", fileId + ".preview"));
+
+							vaultScanResults.incrementNumberOfDeletedContents();
+							vaultScanResults.appendMessage("INFO: Successfully deleted file " + file.getName() + "\n");
+						} catch (Exception e) {
+							vaultScanResults.appendMessage("ERROR: Could not delete file " + file.getName() + "\n");
+							vaultScanResults.appendMessage(e.getMessage() + "\n");
+						}
 					}
 				}
 			}
